+++ History.md
@@ -10,6 +10,7 @@ Dalli Changelog
   functionality already in Dalli.  One benchmark went from 4.0 sec to 3.0
   sec with the new DalliStore.
 - Added reset\_stats operation [#155]
+- Added support for configuring keepalive on TCP connections to memcached servers (@bianster, #180)
 
 Notes:
 
@@ -100,7 +101,7 @@ v1.1.0 was a bad release.  Yanked.
  - Allow browser session cookies (blindsey)
  - Compatibility fixes (mwynholds)
  - Add backwards compatibility module for memcache-client, require 'dalli/memcache-client'.  It makes
-   Dalli more compatible with memcache-client and prints out a warning any time you do something that 
+   Dalli more compatible with memcache-client and prints out a warning any time you do something that
    is no longer supported so you can fix your code.
 
 1.0.1